server/requirements.txt
django
numpy
scipy
scikit-learn
django-registration-redux
djangorestframework
django-rest-auth
django-allauth
celery
django-celery
